requirements.txt

requests
pandas
matplotlib
textblob
nltk
python-dotenv
pytest
